# DRISE GPU optimizations for faster execution time during mask generation